pyproject.toml
[project]
name = "mcp-aurai-advisor"
version = "0.1.1"
description = "上级顾问 MCP 服务器 - 让本地AI获取远程AI指导"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "zhipuai>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-aurai-advisor = "mcp_aurai.server:main"

[project.optional-dependencies]
gui = ["pyinstaller>=6.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]
all-dev = ["mcp-aurai-advisor[gui,dev]"]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_aurai"]

[tool.gui]
config_tool = "tools/config_gui.py"
